+++ admin/views/todosManga.php
@@ -14,39 +14,46 @@
 
 </style>
 
-<h1 class="text-center my-4">Manga</h1>
+<?php if( isset($_SESSION["login"] )) { ?>
 
-<div class="container">  
-    <div class="row row-cols-3">
-        <?php foreach ($mangas as $producto){ ?>
-            <div class="col">
-                <div class="card mb-3" style="max-width: 540px;">
-                    <div class="row g-0">
-                        <div class="col-md-4">
-                            <img src="../img/portadas/<?=$producto->getPortada()?>" class="img-fluid rounded-start" alt="<?= $producto->getTextoAlt() ?>">
-                        </div>
+    <h1 class="text-center my-4">Manga</h1>
 
-                        <div class="col align-self-center">
-                            <div class="card-body">
-                                <h5 class="card-title"><?=$producto->getTitulo()?></h5>
-                                <p class="card-text"><?=$producto->getAutorID()?></p>
-                                <p class="card-text">Volumen: <?=$producto->getVolumen() ?></p>
-                                
-                                <a href="index.php?sec=editar-manga&id=<?= $producto->getID()?>" class="btn add">
-                                    editar
-                                </a>
+    <div class="container">  
+        <div class="row row-cols-3">
+            
+            <?php foreach ($mangas as $producto){ ?>
+                <div class="col">
+                    <div class="card mb-3" style="max-width: 540px;">
+                        <div class="row g-0">
+                            <div class="col-md-4">
+                                <img src="../img/portadas/<?=$producto->getPortada()?>" class="img-fluid rounded-start" alt="<?= $producto->getTextoAlt() ?>">
+                            </div>
 
-                                <a href="index.php?sec=eliminar-Comic&id=<?= $producto->getID()?>" class="btn elim">
-                                    eliminar
-                                </a>
-                                
-                                </button>
+                            <div class="col align-self-center">
+                                <div class="card-body">
+                                    <h5 class="card-title"><?=$producto->getTitulo()?></h5>
+                                    <p class="card-text"><?=$producto->getAutorID()?></p>
+                                    <p class="card-text">Volumen: <?=$producto->getVolumen() ?></p>
+                                    
+                                    <a href="index.php?sec=editar-manga&id=<?= $producto->getID()?>" class="btn add">
+                                        editar
+                                    </a>
+
+                                    <a href="index.php?sec=eliminar-Comic&id=<?= $producto->getID()?>" class="btn elim">
+                                        eliminar
+                                    </a>
+                                    
+                                    </button>
+                                </div>
                             </div>
                         </div>
                     </div>
                 </div>
-            </div>
-        <?php }?>
+            <?php }?>
 
+        </div>
     </div>
-</div>
+<?php }else{
+    header("Location: index.php?sec=dashboard");   
+}
+?>
